update reset password agar sesuai dengan user yang dipih, bukan user yang aktif.

+++ SBR/frontend/controllers/ManageUserController.php
@@ -5,12 +5,14 @@ namespace frontend\controllers;
 use Yii;
 use common\models\User;
 use frontend\models\SearchUser;
-use frontend\models\EditPasswordForm;
-use frontend\models\EditRoleForm;
+use frontend\models\RoleForm;
 use yii\web\Controller;
 use yii\web\NotFoundHttpException;
 use yii\filters\VerbFilter;
 use yii\filters\AccessControl;
+use frontend\models\PasswordResetRequestForm;
+
+//use frontend\models\ResetPasswordForm;
 
 /**
  * UserEditController implements the CRUD actions for User model.
@@ -24,15 +26,15 @@ class ManageUserController extends Controller {
         return [
             'access' => [
                 'class' => AccessControl::className(),
-                'only' => ['index','view', 'update', 'editPassword','editRole'],
-                'rules' => [ 
+                'only' => ['index', 'view', 'update', 'editPassword', 'delete'],
+                'rules' => [
                     [
-                        'actions' => ['view','update','editPassword'],
+                        'actions' => ['view', 'update', 'editPassword'],
                         'allow' => true,
                         'roles' => ['@'],
                     ],
                     [
-                        'actions' => ['editRole','index'],
+                        'actions' => ['editRole', 'index', 'delete'],
                         'allow' => true,
                         'roles' => ['manageUsers'],
                     ],
@@ -46,19 +48,24 @@ class ManageUserController extends Controller {
             ],
         ];
     }
-      
+
     /**
      * Lists all User models.
      * @return mixed
      */
     public function actionIndex() {
-        $searchModel = new SearchUser();
-        $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
+        if (Yii::$app->user->can('manageUsers')) {
+            $searchModel = new SearchUser();
+            $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
 
-        return $this->render('index', [
-                    'searchModel' => $searchModel,
-                    'dataProvider' => $dataProvider,
-        ]);
+            return $this->render('index', [
+                        'searchModel' => $searchModel,
+                        'dataProvider' => $dataProvider,
+            ]);
+        } else {
+            Yii::$app->session->setFlash('warning', 'You do not have permission to access this page.');
+            return $this->goHome();
+        }
     }
 
     /**
@@ -80,47 +87,20 @@ class ManageUserController extends Controller {
      */
     public function actionUpdate($id) {
         $model = $this->findModel($id);
-       $idSuppose = \Yii::$app->user->identity->id;
-              if ($model->id == $idSuppose||\Yii::$app->user->can('manageUsers')) {
+
+        $idSuppose = \Yii::$app->user->identity->id;
+        if ($model->id == $idSuppose || \Yii::$app->user->can('manageUsers')) {
             if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            Yii::$app->session->setFlash('success', 'You have updated this profile.');
-            return $this->redirect(['update', 'id' => $model->id]);
+                Yii::$app->session->setFlash('success', 'You have updated this profile.');
+                return $this->redirect(['view', 'id' => $model->id]);
+            } else {
+                return $this->render('update', [
+                            'model' => $model,
+                ]);
+            }
         } else {
-            return $this->render('update', [
-                        'model' => $model,
-            ]);
-        }
-        }else{
-            Yii::$app->session->setFlash('danger', 'You can not update another user profile.');
+            Yii::$app->session->setFlash('warning', 'You do not have permission to access this page.');
             return $this->goHome();
-        }
-        
-    }
-
-    public function actionEditPassword($id) {
-        $model = new EditPasswordForm($id);
-
-        if ($model->load(Yii::$app->request->post()) && $model->saveNewPassword()) {
-            Yii::$app->session->setFlash('success', 'Password changed.');
-            return $this->redirect(['view', 'id' => $id]);
-        } else {
-            return $this->render('update', [
-                        'model' => $this->findModel($id),
-            ]);
-        }
-    }
-
-    public function actionEditRole($id) {
-        $model = new EditRoleForm($id);
-        //$model = $this->findModel($id);
-
-        if ($model->load(Yii::$app->request->post()) && $model->changeRole()) {
-            Yii::$app->session->setFlash('success', 'Role changed.');
-            return $this->redirect(['view', 'id' => $id]);
-        } else {
-            return $this->render('update', [
-                        'model' => $this->findModel($id),
-            ]);
         }
     }
 
@@ -131,9 +111,14 @@ class ManageUserController extends Controller {
      * @return mixed
      */
     public function actionDelete($id) {
-        $this->findModel($id)->delete();
+        if (Yii::$app->user->can('manageUsers')) {
+            $this->findModel($id)->delete();
 
-        return $this->redirect(['index']);
+            return $this->redirect(['index']);
+        } else {
+            Yii::$app->session->setFlash('warning', 'You do not have permission to access this page.');
+            return $this->goHome();
+        }
     }
 
     /**
@@ -148,6 +133,35 @@ class ManageUserController extends Controller {
             return $model;
         } else {
             throw new NotFoundHttpException('The requested page does not exist.');
+        }
+    }
+
+//    public function actionChangePassword() {
+//        $model = new PasswordResetRequestForm();
+//        $model->email = Yii::$app->user->identity->email;
+//        if ($model->validate()) {
+//            if ($model->sendEmail()) {
+//                Yii::$app->session->setFlash('success', 'Check your email for further instructions.');
+//
+//                return $this->goHome();
+//            } else {
+//                Yii::$app->session->setFlash('error', 'Sorry, we are unable to reset password for email provided.');
+//            }
+//        }
+//    }
+    public function actionChangePassword($id) {//menggunakan reset password form
+        $model = new PasswordResetRequestForm();
+
+        $model->email = $this->findModel($id)->email; //email dari user yang dipilih, bukan user yang aktif.
+        //$model->email = Yii::$app->user->identity->email; //email dari user yang aktif
+        if ($model->validate()) {
+            if ($model->sendEmail()) {
+                Yii::$app->session->setFlash('success', 'Check your email for further instructions.');
+
+                return $this->goHome();
+            } else {
+                Yii::$app->session->setFlash('error', 'Sorry, we are unable to reset password for email provided.');
+            }
         }
     }
